Move trait definitions to their dedicated files
